requirements/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
django-cors-headers==3.14.0
requests
mock
spacy
